Main pipeline is now the "RouterService". WaxHub should work in single-threaded mode when running in DEBUG
