+++ Compiler/CommonUtil/Wax/WaxHub.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace CommonUtil.Wax
 {
@@ -17,19 +18,137 @@ namespace CommonUtil.Wax
             Dictionary<string, object> request,
             Func<Dictionary<string, object>, bool> cb)
         {
+            if (cb == null) cb = d => false;
+
             WaxService service = this.services.ContainsKey(serviceName) ? this.services[serviceName] : null;
             if (service == null)
             {
                 // TODO: check for extensible services online
                 throw new Exception("Invalid service name: '" + serviceName + "'");
             }
-            string encodedRequest = WaxService.SerializeWireData(request);
+            string encodedRequest = SerializeWireData(request);
             System.ComponentModel.BackgroundWorker bg = new System.ComponentModel.BackgroundWorker();
             bg.DoWork += (e, sender) =>
             {
-                service.HandleRequest(WaxService.ParseWireData(encodedRequest), cb);
+                service.HandleRequest(ParseWireData(encodedRequest), cb);
             };
             bg.RunWorkerAsync();
+        }
+
+        public Dictionary<string, object> AwaitSendRequest(
+            string serviceName,
+            Dictionary<string, object> request)
+        {
+            List<Dictionary<string, object>> responsePtr = new List<Dictionary<string, object>>();
+#if DEBUG
+            this.services[serviceName].HandleRequest(
+                ParseWireData(SerializeWireData(request)),
+                response => {
+                    responsePtr.Add(response);
+                    return true;
+                });
+
+            if (responsePtr.Count == 0) throw new NotImplementedException();
+            return responsePtr[0];
+#else
+
+            this.SendRequest(serviceName, request, response =>
+            {
+                lock (responsePtr)
+                {
+                    responsePtr.Add(response);
+                }
+                return true;
+            });
+
+            double counter = 1;
+            while (true)
+            {
+                lock (responsePtr)
+                {
+                    if (responsePtr.Count > 0)
+                    {
+                        return responsePtr[0];
+                    }
+                }
+                System.Threading.Thread.Sleep((int)counter);
+                counter *= 1.5;
+                if (counter > 100) counter = 100;
+            }
+#endif
+        }
+
+
+        internal static string SerializeWireData(Dictionary<string, object> data)
+        {
+            List<string> wireData = new List<string>();
+            foreach (string key in data.Keys)
+            {
+                object value = data[key];
+                wireData.Add(key);
+                if (value is int)
+                {
+                    wireData.Add("I");
+                    wireData.Add("" + data);
+                }
+                else if (value is bool)
+                {
+                    wireData.Add("B");
+                    wireData.Add((bool)value ? "1" : "0");
+                }
+                else if (value is string)
+                {
+                    wireData.Add("S");
+                    wireData.Add((string)value);
+                }
+                else if (value is string[])
+                {
+                    wireData.Add("A");
+                    string[] items = (string[])value;
+                    wireData.Add(items.Length + "");
+                    foreach (string item in items)
+                    {
+                        wireData.Add(item);
+                    }
+                }
+                else
+                {
+                    throw new NotImplementedException();
+                }
+            }
+
+            return string.Join(',', wireData.Select(item => Base64.ToBase64(item)));
+        }
+
+        internal static Dictionary<string, object> ParseWireData(string encodedData)
+        {
+            Dictionary<string, object> output = new Dictionary<string, object>();
+            string[] items = encodedData.Split(',').Select(item => Base64.FromBase64(item)).ToArray();
+            for (int i = 0; i < items.Length; i += 3)
+            {
+                string key = items[i];
+                char type = items[i + 1][0];
+                string data = items[i + 2];
+                switch (type)
+                {
+                    case 'B': output[key] = data == "1"; break;
+                    case 'I': output[key] = int.Parse(data); break;
+                    case 'S': output[key] = data; break;
+
+                    case 'A':
+                        List<string> arr = new List<string>();
+                        int length = int.Parse(data);
+                        for (int j = 0; j < length; ++j)
+                        {
+                            arr.Add(items[i++ + 3]);
+                        }
+                        output[key] = arr.ToArray();
+                        break;
+
+                    default: throw new NotImplementedException();
+                }
+            }
+            return output;
         }
     }
 }
